requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-surprise
sqlalchemy
pymysql
joblib
pytest
python-dotenv
statsmodels
